fix: move space metadataheader to spaces-page partials
